Create SQL project  on census of India
The data used is not the whole census of India. A part of data which contains 1000 rows and 640 rows in respective datasheets is taken for the project.

The project includes sql commands from the very basic sleecting columns to joining tables.

The project is made from the reference and guidance of youtube videos.